完善了初始化接口

+++ GYCS-Admin/gycs-charity/src/main/java/com/ruoyi/charity/domain/dto/CharityRaiseFund.java
@@ -63,7 +63,7 @@ public class CharityRaiseFund
     /** 活动状态 */
     @Excel(name = "活动状态")
     @TableField(value = "status")
-    private Long status;
+    private BigInteger status;
